feat: if 'if state' = true now outputs to the first output
For current-state, poll-state and event-state nodes: when the 'if state' is true the node will now send the msg to the first output.
Existing nodes will continue to function in the old manner with a true 'if state' being sent to the second output until the node is edited in the UI and automatically updated to use the new way.

Nodes now have a version number assigned to them.