feat(trigger-state): Update listener to allow listening to multiple types at once
This commit updates the error message in the `trigger-state` module to provide a more descriptive explanation when an entity is required but not provided. The code changes in `trigger-state/locale.json` modify the `entity_id_required` error message to `entity_required`. This improvement enhances the clarity and usability of the `trigger-state` node.

#fixes: 1103
#fixes: 1402